--- packages/builder/src/pages/builder/app/[application]/automation/[automationId]/_flagged/layout.old.svelte
@@ -1,15 +1,0 @@
-<script>
-  import { automationStore } from "@/stores/builder"
-  import { params } from "@roxi/routify"
-
-  if ($params.automation) {
-    const automation = $automationStore.automations.find(
-      m => m._id === $params.automation
-    )
-    if (automation) {
-      automationStore.actions.select(automation)
-    }
-  }
-</script>
-
-<slot />
